+++ src/utils/constants.js
@@ -118,6 +118,7 @@ export const ENTRY_HEIGHT = 40;
 export const ELEVATION_DEFAULT = 45;
 
 export const FOOTER_HEIGHT = 20;
+export const FOOTER_SPACER = 130;
 
 export const FREQUENCY_GHZ = 1;
 export const FREQUENCY_MHZ = 2;
@@ -181,9 +182,9 @@ export const IMAGE_WEIGHTING = [
 ];
 
 export const IMAGE_SIZE_UNITS = {
-  ARCSECS: 'arcseconds',
-  ARCMINS: 'arcmins',
-  DEGREES: 'degrees'
+  ARCSECS: 'arcsec',
+  ARCMINS: 'arcmin',
+  DEGREES: 'degree'
 };
 
 export const LAB_IS_BOLD = true;
@@ -846,6 +847,8 @@ export const VELOCITY_UNITS = [
   { label: 'm/s', value: 0, convert: 1 },
   { label: 'km/s', value: 1, convert: 1000 }
 ];
+
+export const WRAPPER_HEIGHT = '80px';
 
 export const LOW_BEAM_SIZE_PRECISION = 1;
 export const MID_BEAM_SIZE_PRECISION = 3;