+++ src/utils/types/metadata.tsx
@@ -2,6 +2,7 @@ export type Metadata = {
   version: number;
   created_by: string;
   created_on: string;
+  pdm_version: string;
   last_modified_by: string;
   last_modified_on: string;
 };